Updated criteria for detecting death in cavebot and autolooter(hpPercLeft no longer set to 0 and only visible=0 and onscreenMax changes to maxint) Fixed crash when using %d with 64bit time_t instead of %lld

+++ tibiaauto-kernel/PythonEngine.cpp
@@ -907,7 +907,7 @@ void CPythonEngine::backpipeInpacketTick()
 		}
 		else     //everything is OK
 		{
-			sprintf(buf, "One packet took %dms", GetTickCount() - tm);
+			sprintf(buf, "One packet took %dms", (int)(GetTickCount() - tm));
 			//if (GetTickCount() - tm>20) MessageBox(NULL, buf,"huaeotnre",0);
 			CPythonScript *pythonScript = NULL;
 			for (int scriptNr = 0; (pythonScript = CPythonScript::getScriptByNr(scriptNr)); scriptNr++)
